Fix different Fuzz tests issues
This commit fixes several issues found during fuzz tests:
- adress sanitizer issue in MCTP control demon in
	mctp_i2c_get_msg_type_response function;
- issue with proper synchronization with SPI thread in
	MCTP control daemon;
- issue with direct memory leak in MCTP demux daemon
	(missing free usage after realloc);
- issue with improper support for too short messages
	(missing condition check in binding function);
- issue with usage of a memory after free call;
- issue with broken functionality for smbus bridge interface
	type;
- issue with usage of uninitialized buffer in smbus.c
	when sending get udid message;
- issue with hang up in the main demux daemon loop,
	when a client disconnects before providing its type.

The changes were verified on ranger and hgx devices:
- running all mctp services: smbus, pcie, spi;
- checking initial messages and journal logs for any
	unexpected behaviour;
- creating required mctp ctrl sdbus objects;
- introspecting MCTP dbus objects and compared with
	an image without this commit changes;
- checking vdm util on all available mctp objects.

Fixes nvbug https://nvbugspro.nvidia.com/bug/4400879
Fixes nvbug https://nvbugspro.nvidia.com/bug/4404363
Fixes nvbug https://nvbugspro.nvidia.com/bug/4466081
Fixes nvbug https://nvbugspro.nvidia.com/bug/4466555
Fixes nvbug https://nvbugspro.nvidia.com/bug/4466962
Fixes nvbug https://nvbugspro.nvidia.com/bug/4493075
Fixes nvbug https://nvbugspro.nvidia.com/bug/4502094
Fixes nvbug https://nvbugspro.nvidia.com/bug/4509770

+++ ctrld/mctp-discovery-common.h
@@ -53,7 +53,7 @@ typedef struct mctp_routing_table {
 /* List for MCTP Message types */
 typedef struct mctp_msg_type_table {
 	uint8_t eid;
-	int data_len;
+	uint16_t data_len;
 	uint8_t data[MCTP_MSG_TYPE_MAX_SIZE];
 	struct mctp_msg_type_table *next;
 } mctp_msg_type_table_t;
